支付功能 加入支付宝沙箱 秒杀服务，服务上架功能完成，并测试

+++ gulimall-coupon/src/main/java/com/atguigu/gulimall/coupon/entity/SeckillSkuRelationEntity.java
@@ -44,11 +44,11 @@ public class SeckillSkuRelationEntity implements Serializable {
 	/**
 	 * ??ɱ????
 	 */
-	private BigDecimal seckillCount;
+	private Integer seckillCount;
 	/**
 	 * ÿ???޹??
 	 */
-	private BigDecimal seckillLimit;
+	private Integer seckillLimit;
 	/**
 	 * ???
 	 */
